Resolve #329: integrate CloudKit JS as alternate web backend
Enables the previously-disabled "CloudKit JS (browser)" toggle in the
mistdemo web UI and wires real container.privateCloudDatabase calls
(performQuery / saveRecords / deleteRecords) behind it, so each CRUD
operation can be driven either through MistKit on the server or
CloudKit JS in the browser against the same container with the same
auth token.

- index.html: replace cloudKitJsNotWired() short-circuits with real
  CloudKit JS handlers; add fieldsToCKJS() to wrap the demo's flat
  {key: "value"} shape into CloudKit JS's {key: {value: ...}} form;
  add runCloudKitJsOperation() to surface the hasErrors/errors payload
  pattern that CloudKit JS uses in place of throwing.
- Update mode hint: both backends share the same Apple ID session
  token, container, and REST surface; only the SDK shape differs.
- Move the index-page tests into WebServerTests+Index.swift (matching
  the existing +CRUD extension pattern) and add a smoke test asserting
  the CDN script tag, an enabled toggle, the three call sites, and
  absence of the placeholder.

No server-side changes: the auth-token capture flow already shares
the ckWebAuthToken across both modes via WebAuthTokenStore.
