+++ Examples/MistDemo/Tests/MistDemoTests/Server/WebServerTests.swift
@@ -71,20 +71,6 @@
       return Fixture(server: server, tokenStore: store, backend: backend)
     }
 
-    @Test("GET / returns the web demo HTML")
-    internal func indexReturnsHtml() async throws {
-      let fixture = Self.makeFixture()
-      let app = Application(router: try fixture.server.makeRouter())
-
-      try await app.test(.router) { client in
-        try await client.execute(uri: "/", method: .get) { response in
-          #expect(response.status == .ok)
-          let body = String(buffer: response.body)
-          #expect(body.contains("MistKit Web Demo"))
-        }
-      }
-    }
-
     @Test("GET /api/config returns container + environment")
     internal func configIncludesEnvironment() async throws {
       let fixture = Self.makeFixture()
